+++ HideousGameEngine/test/TestTemplate.h
@@ -16,7 +16,7 @@
 class ITest{
 public:
 	virtual ~ITest(){}
-	virtual void Update(double dt) = 0;
+	virtual void Update(float dt) = 0;
 	virtual void Render() = 0;
 
 };
